+++ metadata.rb
@@ -4,6 +4,6 @@ maintainer_email 'ainfamoussyn@gmail.com'
 license          'All rights reserved'
 description      'Installs/Configures CTF-Cookbook'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '2.0.0'
+version          '3.0.1'
 
 depends "apt", ">= 2.7.0"