+++ recipes/ctf-tools.rb
@@ -8,7 +8,7 @@
 #
 
 cookbook_file 'checksec.sh' do
-	path "#{node['general']['dir']}/checksec.sh"
+	path "#{node['general']['dir']}/#{node['files']['checksec']}"
 	owner node['general']['owner']
 	group node['general']['group']
 	mode node['general']['mode']
@@ -16,7 +16,7 @@ cookbook_file 'checksec.sh' do
 end
 
 cookbook_file 'shellcode_examiner.c' do
-	path "#{node['general']['dir']}/shellcode_examiner.c"
+	path "#{node['general']['dir']}/#{node['files']['shellcode_examiner']}"
 	owner node['general']['owner']
 	group node['general']['group']
 	mode node['general']['mode']
@@ -25,74 +25,32 @@ end
 
 # Installs required dependencies and setup IDA Demo.
 if node['ida']['enabled']
-	cookbook_file "#{node['ida']['package']}" do
-		path "#{node['general']['dir']}/#{node['ida']['package']}"
-		owner node['general']['owner']
-		group node['general']['group']
-		mode node['general']['mode']
-		action :create_if_missing
-	end
-
-	execute 'Unpacking IDA archive' do
-		cwd "#{node['general']['dir']}"
-		command "tar zxf #{node['general']['dir']}/#{node['ida']['package']}"
-	end
+	include_recipe 'CTF-Cookbook::ctf-tools-ida'
 end
 
 # Installs required dependencies and setup Hopper Debugger.
 if node['hopper']['enabled']
 	include_recipe 'apt::default'
+	include_recipe 'CTF-Cookbook::ctf-tools-hopper'
+end
 
-	node[:hopper][:depends].each do |hopper_depends|
-		package hopper_depends do
-			action :install
-		end
- 	end
+# Installs the rardare2 tool
+if node['radare2']['enabled']
+	include_recipe 'CTF-Cookbook::ctf-tools-radare2'
+end
 
-	cookbook_file "#{node['hopper']['package']}" do
-		path "#{node['general']['dir']}/#{node['hopper']['package']}"
-		action :create_if_missing
-	end
+if node['pwntools']['enabled']
+	include_recipe 'CTF-Cookbook::ctf-tools-pwntools'
+end
 
-	execute 'Installing Hopper debugger' do
-		command "dpkg -i #{node['general']['dir']}/#{node['hopper']['package']}"
-	end
-
-	link "#{node['hopper']['dir']}/hopper-v3/hopper-launcher.sh" do
-		to '/usr/bin/hopper'
-	end
+if node['qemu']['enabled']
+	include_recipe 'CTF-Cookbook::ctf-tools-qemu'
 end
 
 # Creates the Burp .jar file.
 if node['java']['enabled']
 	if node['burp']['enabled']
-		cookbook_file "#{node['burp']['package']}" do
-			path "#{node['general']['dir']}/#{node['burp']['package']}"
-			owner node['general']['owner']
-			group node['general']['group']
-			mode node['general']['mode']
-			action :create_if_missing
-		end
+		include_recipe 'CTF-Cookbook::ctf-tools-burp'
 	end
 end
 
-if node['pwntools']['enabled']
-	node[:pwntools][:depends].each do |pwntools_depends|
-		package pwntools_depends do
-			action :install
-		end
-	end
-
-	# https://pwntools.readthedocs.org/en/latest/install.html
-	execute 'Installing Pwntools' do
-		command "pip install #{node['pwntools']['package']}"
-	end
-end
-
-if node['qemu']['enabled']
-	node[:qemu][:version].each do |qemu_version|
-		package qemu_version do
-			action :install
-		end
-	end
-end